+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated examples and tests for `terra-table` to consume udpated onRowSelect callback.
+
 ## 1.49.0 - (December 1, 2023)
 
 * Added